requirements.txt
#requirement placeholder
kafka-python==1.4.4
requests==2.21.0
objectpath==0.6.1
kubernetes==8.0.1